Removed old, renamed v2 test_aggregation.
